+++ js/remote-work.js
@@ -1,43 +1,40 @@
-import { showAlert } from './util.js';
-import { showErrorSend, showSuccessSend } from './util.js';
+import {showErrorSend, showSuccessSend} from './modal/AjaxModals.js';
+import {showAlert} from './alerts/send-alert.js';
 
-const getData = (onSuccess) => {
-  fetch('https://24.javascript.pages.academy/kekstagram/data')
+const myFetch = (url, options, onSuccess, onFailed) => {
+  return fetch(url, options)
     .then((response) => {
       if (response.ok) {
-        return response.json();
+        onSuccess && onSuccess();
+        return response;
       }
-      throw new Error(`${response.status} ${response.statusText}`);
+      throw new Error(response.statusText);
     })
-    .then((data) => {
-      onSuccess(data);
-    }).catch((err) => {
-      showAlert(`При загрузке данных произошла ошибка \n ${err}`);
+    .catch(() => {
+      onFailed && onFailed();
     });
+};
+
+const getData = () => {
+  const url = 'https://24.javascript.pages.academy/kekstagram/data';
+  const options = {};
+  const failedFunction = () => showAlert('При загрузке данных произошла ошибка');
+  return myFetch(url, options, null, failedFunction)
+    .then((response) => response.json);
 };
 
 const sendData = (post) => {
-  fetch(
-    'https://24.javascript.pages.academy/kekstagram',
-    {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'multipart/form-data',
-      },
-      body: JSON.stringify(post),
+  const url = 'https://24.javascript.pages.academy/kekstagram';
+  const options = {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'multipart/form-data',
     },
-  )
-    .then((response) => {
-      if (response.ok) {
-        showSuccessSend();
-      } else {
-        showErrorSend();
-      }
-    })
-    .catch(() => {
-      showErrorSend();
-    });
+    body: JSON.stringify(post),
+  };
+  const FailedFunction = showErrorSend;
+  const SuccessFunction = showSuccessSend;
+  return myFetch(url, options, SuccessFunction, FailedFunction)
 };
-
 
 export {getData, sendData};
